refactor(skeletonlistitemrow.tsx): separate component for skeleton list item
